requirements.txt
django==5.2.5
sqlparse==0.5.3
typing-extensions==4.13.2
black==24.8.0
click==8.2.1
django==5.2.5
djangorestframework==3.16.1
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
sqlparse==0.5.3
tomli==2.2.1
typing-extensions==4.13.2
